# api/webapi/blog.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "blog.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BlogService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/Blog": {
      "get": {
        "operationId": "BlogService_GetBlog",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/BlogResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "delete": {
        "operationId": "BlogService_DeleteBlog",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/BlogResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "post": {
        "operationId": "BlogService_CreateBlog",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateBlogResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "blog.id",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "blog.createdAt",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "blog.updatedAt",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "blog.imagePath",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "blog.title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "blog.body",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "blog.version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "put": {
        "operationId": "BlogService_UpdateBlog",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/BlogResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/Blogs": {
      "get": {
        "operationId": "BlogService_GetAllBlogs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetAllBlogsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    }
  },
  "definitions": {
    "Blog": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "imagePath": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
    "BlogResponse": {
      "type": "object",
      "properties": {
        "blog": {
          "$ref": "#/definitions/Blog"
        }
      }
    },
    "CreateBlogResponse": {
      "type": "object",
      "properties": {
        "blog": {
          "$ref": "#/definitions/Blog"
        }
      }
    },
    "GetAllBlogsResponse": {
      "type": "object",
      "properties": {
        "blog": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Blog"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}